feat: maintain adjacent comment alignment
Previously a block of comments that had been manually aligned vertically
over consecutive lines like this:

    Op == /\ bananas=42         \* This is an important number.
          /\ platanos'  =  42   \* That should be assigned here.

Would each be formatted in isolation, leading to:

    Op == /\ bananas = 42 \* This is an important number.
          /\ platanos' = 42 \* That should be assigned here.

Which obviously isn't the authors intent, and can harm readability.
After this commit, the alignment and positioning are retained, even if
the line lengths change due to formatting changes:

    Op == /\ bananas = 42       \* This is an important number.
          /\ platanos' = 42     \* That should be assigned here.

If a line is formatted causing its length to grow past the original
comment position, the whole block is pushed back to maintain alignment.

+++ libtlafmt/src/helpers/indent.rs
@@ -10,7 +10,7 @@ pub(crate) struct IndentDecorator<W> {
     out: W,
 
     /// The indentation depth to render for subsequent writes.
-    depth: usize,
+    depth: u8,
 
     /// True when the last byte wrote to `out` was a newline.
     last_char_newline: bool,
@@ -25,7 +25,7 @@ impl<W> IndentDecorator<W> {
         }
     }
 
-    pub(crate) fn set(&mut self, depth: usize) {
+    pub(crate) fn set(&mut self, depth: u8) {
         self.depth = depth;
     }
 }
